implementation/src/requirements.txt
Pillow == 9.1.0
matplotlib == 3.4.3
numpy == 1.19.4
scikit_image == 0.19.2
scipy == 1.7.1
